% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maria.R
\name{delete_from_table}
\alias{delete_from_table}
\title{Delete query}
\usage{
delete_from_table(
  table_name_in_base,
  where,
  host = "localhost",
  port = 3306,
  db,
  user,
  password
)
}
\arguments{
\item{table_name_in_base}{table in \code{db} to delete rows from}

\item{where}{SQL WHERE clause (without the WHERE keyword) selecting rows to delete. Interpolated verbatim into the statement -- the caller is responsible for sanitizing any untrusted input (this fragment is NOT escaped).}

\item{host}{host}

\item{port}{port}

\item{db}{default database name}

\item{user}{user}

\item{password}{password}
}
\description{
Delete from table rows that match certain criteria
}
\examples{
\dontrun{delete_from_table(table_name_in_base="foo", where="id in (1, 2, 3)", host=HOST, db=DB, user=USER, password=PWD)}
}
\keyword{delete}
\keyword{mysql}
