% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{execq}
\alias{execq}
\title{Exec}
\usage{
execq(query, ...)
}
\arguments{
\item{query}{query to execute}

\item{...}{any other argument passed to \code{exec_query}}
}
\description{
Simple wrapper around `exec_query` method, that makes credential use transparent.
Requires credentials to be loaded, obviously.
}
\examples{
\dontrun{execq("TRUNCATE TABLE foo;")}
}
\seealso{
exec_query pull_data
}
\keyword{mysql}
\keyword{select}
