% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maria.R
\name{insertq}
\alias{insertq}
\title{Simplified insert}
\usage{
insertq(table, table_name_in_base, ...)
}
\arguments{
\item{table}{data.frame or data.table to insert}

\item{table_name_in_base}{table in {db} to insert data into}

\item{...}{any other parameter that applies to insert_table}
}
\description{
Simple method that inserts the input data.frame or data.table into the designated table in the current DB context.
}
\details{
It's important to be aware that both input table and table in database should have the same schema (matching names, matching types).
}
\examples{
\dontrun{data <- insertq(host=HOST, db=DB, user=user, password=pwd, query="select * from table;")}
}
\seealso{
pull_data, selectq, insert_table
}
\keyword{insert}
\keyword{mysql}
