% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maria.R
\name{selectq}
\alias{selectq}
\title{Select}
\usage{
selectq(query, ...)
}
\arguments{
\item{query}{query to execute}

\item{...}{any argument that can be sent to `pull_data`}
}
\description{
Simple wrapper around `pull_data` method, that makes credential use transparent.
Requires credentials to be loaded, obviously.
}
\examples{
\dontrun{load_env("dwh", 123)
selectq("select * from shop limit 10;")}
}
\seealso{
insert_query pull_data
}
\keyword{mysql}
\keyword{select}
