% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maria.R
\name{truncate_table}
\alias{truncate_table}
\title{Truncate table}
\usage{
truncate_table(
  table_name_in_base,
  host = "localhost",
  port = 3306,
  db,
  user,
  password
)
}
\arguments{
\item{table_name_in_base}{table in {db} to insert data into}

\item{host}{host}

\item{port}{port}

\item{db}{default database name}

\item{user}{user}

\item{password}{password}
}
\description{
Empties table from all observations, but doesn't delete it.
}
\examples{
\dontrun{truncate_table(table="foo", host=HOST, db=DB, user=USER, password=PWD)}
}
\keyword{delete}
\keyword{mysql}
