% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maria.R
\name{updateq}
\alias{updateq}
\title{Simplified update}
\usage{
updateq(table, table_name_in_base, ...)
}
\arguments{
\item{table}{data.frame or data.table to update}

\item{table_name_in_base}{table in \code{db} to update data into}

\item{...}{any other parameter that applies to update_table}
}
\description{
Simple method that updates the input data.frame or data.table into the designated table in the current DB context.
}
\details{
It's important to be aware that both input table and table in database should have the same schema (matching names, matching types).
}
\examples{
\dontrun{updateq(iris, "iris_database_name")}
}
\seealso{
pull_data, selectq, update_table, insertq, insert_table
}
\keyword{insert}
\keyword{mysql}
\keyword{update}
