% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maria.R
\name{upsert_table}
\alias{upsert_table}
\title{Upsert}
\usage{
upsert_table(
  table,
  table_name_in_base,
  keycols,
  host = "localhost",
  port = 3306,
  db,
  user,
  password,
  progress_bar = interactive(),
  nolog = FALSE
)
}
\arguments{
\item{table}{data.frame or data.table to insert}

\item{table_name_in_base}{table in \code{db} to upsert data into}

\item{keycols}{character vector naming the key column(s) used to identify rows (excluded from the SET/UPDATE clause)}

\item{host}{host}

\item{port}{port}

\item{db}{default database name}

\item{user}{user}

\item{password}{password}

\item{progress_bar}{nice progress bar to use, it's recommended to disable it in log mode}

\item{nolog}{avoid any writing to the console (when TRUE, errors are not logged either)}
}
\description{
Simple method that inserts the input data.frame or data.table into the designated table, or updates it if the key already exists.
}
\details{
It's important to be aware that both input table and table in database should have the same schema (matching names, matching types).
}
\examples{
\dontrun{upsert_table(my_data, "table_name", keycols=c("id"), host=HOST, db=DB, user=USER, password=PWD)}
}
\seealso{
pull_data, selectq, insertq
}
\keyword{insert}
\keyword{mysql}
