% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maria.R
\name{upsertq}
\alias{upsertq}
\title{Simplified upsert}
\usage{
upsertq(table, table_name_in_base, ...)
}
\arguments{
\item{table}{data.frame or data.table to upsert}

\item{table_name_in_base}{table in {db} to upsert data into}

\item{...}{any other parameter that applies to upsert_table}
}
\description{
Simple method that upserts the input data.frame or data.table into the designated table in the current DB context.
}
\details{
It's important to be aware that both input table and table in database should have the same schema (matching names, matching types).
}
\examples{
\dontrun{upsertq(iris, "iris_database_name")}
}
\seealso{
pull_data, selectq, upsert_table, insertq, insert_table
}
\keyword{insert}
\keyword{mysql}
\keyword{update}
\keyword{upsert}
