% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maria.R
\name{deleteq}
\alias{deleteq}
\title{Simplified delete query}
\usage{
deleteq(table_name_in_base, where, ...)
}
\arguments{
\item{table_name_in_base}{table in \code{db} to insert data into}

\item{where}{SQL where clause (wtihout the keyword where) specifying which rows should be deleted}
}
\description{
Delete from table rows that match certain criteria
}
\examples{
\dontrun{deleteq(table_name_in_base="foo", where="id < 10")}
}
\keyword{delete}
\keyword{mysql}
