% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maria.R
\name{execq}
\alias{execq}
\title{Exec}
\usage{
execq(query, ...)
}
\arguments{
\item{query}{query to execute}

\item{...}{any argument that can be sent to `pull_data`}
}
\description{
Simple wrapper around `exec_query` method, that makes credential use transparent.
Requires credentials to be loaded, obviously.
}
\examples{
\dontrun{execq('set character set "utf8"')}
}
\seealso{
exec_query pull_data
}
\keyword{mysql}
\keyword{select}
