% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maria.R
\name{insert_table_local}
\alias{insert_table_local}
\title{Simplified bulk insert}
\usage{
insert_table_local(
  table,
  table_name_in_base,
  preface_queries = character(0),
  split_threshold = 100000,
  use_file = FALSE
)
}
\arguments{
\item{table}{data.frame or data.table to insert}

\item{table_name_in_base}{table in {db} to insert data into}

\item{preface_queries}{character vector of queries you want to apply before, typically setting session variables.}

\item{split_threshold}{integer, number of rows to split the data into smaller groups. Default is 1e5.}

\item{use_file}{logical, if TRUE, uses `load_data_local_infile` flag. Default is FALSE. Careful if you use it, as it does a count(*) on table before and after to check what was integrated.}
}
\description{
Simple method that inserts the input data.frame or data.table into the designated table in the current DB context.
}
\details{
It's important to be aware that both input table and table in database should have the same schema (matching names, matching types). The difference between insertq and insert_table_local is that \code{insertq} uses homemade INSERTS statements, and \code{insert_table_local} uses `load_data_local_infile` flag.
}
\examples{
\dontrun{
  data <- insert_table_local(iris, "iris")
  data <- insert_table_local(iris, "iris", preface_queries="SET session rocksdb_bulk_load=1")
}
}
\seealso{
pull_data, selectq, insert_table, insertq
}
\keyword{MariaDB}
\keyword{insert}
