% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maria.R
\name{pull_data}
\alias{pull_data}
\title{Detailed select}
\usage{
pull_data(
  host = "localhost",
  port = 3306,
  db,
  user,
  password,
  query,
  scroll_size = NA,
  verbose = TRUE,
  keep_int64 = FALSE
)
}
\arguments{
\item{host}{host}

\item{port}{port}

\item{db}{default database name}

\item{user}{user}

\item{password}{password}

\item{query}{query to execute}

\item{scroll_size}{scroll size}

\item{verbose}{output current state}
}
\description{
Simple method that executes your select query and returns its results in a data.table.
}
\examples{
\dontrun{data <- pull_data(host=HOST, db=DB, user=user, password=pwd, query="select * from table;")}
}
\seealso{
insert_table
}
\keyword{mysql}
\keyword{select}
